Major Refactor to add categories to operators. Used the time to convert to new-style constructor to create builtin operators. Added more unit tests to verify changes.

+++ src/src/org/tms/teq/MathUtil.java
@@ -357,7 +357,7 @@ public class MathUtil
             return Double.NaN;
     }   
     
-    static final public double numberOf(TableElement te, Object q)
+    static final public int numberOf(TableElement te, Object q)
     {
         int count = 0;
         if (te != null && q != null) {
@@ -385,7 +385,7 @@ public class MathUtil
             }
         }
         
-        return (double)count;
+        return count;
     }   
     
     static final private BigDecimal toBigDecimal(Object n)
